Introduce Webview for all course Modules
The api sends weburls for all course content instead of actual content.
Instead of doing something very cumbersome like parsing html and pulling
multiple apis, we can make use of the urls to give a native experience
while displaying content. This also fixes some small bugs.
