+++ EdxMobile/app/screens/index.ts
@@ -3,9 +3,9 @@ export * from "./LoginScreen"
 export * from "./ProfileScreen"
 export * from "./ErrorScreen/ErrorBoundary"
 export * from "./CourseDetailScreen"
-export * from "./ModuleScreen"
-export * from "./DiscussionScreen"
-export * from "./DiscussionThreadScreen"
+export * from "./Modules/ModuleScreen"
+export * from "./Discussion/DiscussionScreen"
+export * from "./Discussion/DiscussionThreadScreen"
 export * from "./SignUpScreen"
 export * from "./ProgressScreen"
 export * from "./DatesScreen/DatesScreen"
