+++ EdxMobile/app/utils/formatData.ts
@@ -36,10 +36,9 @@ export function normalizeOutlineBlocks(courseId, blocks) {
             effortTime: block.effort_time,
             icon: block.icon,
             id: block.id,
-            // The presence of a URL for the sequence indicates that we want this sequence to be a clickable
-            // link in the outline (even though we ignore the given url and use an internal <Link> to ourselves).
             showLink: !!block.lms_web_url,
             title: block.display_name,
+            lms_web_url: block.lms_web_url
           };
           break;
