+++ EdxMobile/app/utils/formatDate.ts
@@ -13,7 +13,6 @@ const getLocale = (): Locale => {
 }
 
 export const formatDate = (date: string, dateFormat?: string, options?: Options) => {
-  console.log(date)
   const locale = getLocale()
   const dateOptions = {
     ...options,
